+++ tests/Integration/Commands/ConvertTo-SqlDscDataFile.Integration.Tests.ps1
@@ -34,7 +34,7 @@ BeforeAll {
     Import-SqlDscPreferredModule
 }
 
-Describe 'ConvertTo-SqlDscDataFile' -Tag @('Integration_SQL2017', 'Integration_SQL2019', 'Integration_SQL2022') {
+Describe 'ConvertTo-SqlDscDataFile' -Tag @('Integration_SQL2017', 'Integration_SQL2019', 'Integration_SQL2022', 'Integration_SQL2025') {
     BeforeAll {
         $script:mockInstanceName = 'DSCSQLTEST'
         $script:mockComputerName = Get-ComputerName