+++ tests/Integration/Commands/Initialize-SqlDscImage.Integration.Tests.ps1
@@ -32,7 +32,7 @@ BeforeAll {
 }
 
 # cSpell: ignore SQLSERVERAGENT, DSCSQLTEST
-Describe 'Initialize-SqlDscImage' -Tag @('Integration_SQL2017', 'Integration_SQL2019', 'Integration_SQL2022') {
+Describe 'Initialize-SqlDscImage' -Tag @('Integration_SQL2017', 'Integration_SQL2019', 'Integration_SQL2022', 'Integration_SQL2025') {
     BeforeAll {
         $computerName = Get-ComputerName
         Write-Verbose -Message ("Running integration test as user '{0}' on computer '{1}'." -f $env:UserName, $computerName) -Verbose